fix normalisation in training and prediction
Reviewing my code I've realised that I made a mistake in my
implementation of the normalisation of the triplet image sets. Although
the normalisation functions were being called the result of their
operation was not actually being returned (in
`DeviceDataLoader.proc_batch`). This commit fixes that and seems to
improve training quite a lot.