Bugfix for producing embeddings from triplets
I hadn't fully tested the changes made to support fastai `1.0.52`. This
commit fixes `get_encodings` (should be renamed `get_embeddings`) so
that it correctly calls the trained models `predict` function. We also
add two `reconstruct` methods for the input (triplet) and output (empty
label) of the model as fastai uses these to attempt to "reconstruct" an
output image from the input.