dendrogram plotting in single figure
Dendrogram plot code has been refactored so that all components of plot
are inside the figure's [0:1]x[0:1] extent. This ensure that tiles are
clipped off when uploading to W&B. The refactor also makes it easier in
future to put dendrogram annotations on other kinds of plots (for
example confusion matrices).